Move the tests to a separate file
